+++ src/SystemCare/Views/BatteryHealthPage.xaml.cs
@@ -1,0 +1,19 @@
+using System.Windows;
+using System.Windows.Controls;
+using SystemCare.ViewModels;
+
+namespace SystemCare.Views;
+
+public partial class BatteryHealthPage : Page
+{
+    private readonly BatteryHealthViewModel _viewModel;
+
+    public BatteryHealthPage(BatteryHealthViewModel viewModel)
+    {
+        _viewModel = viewModel;
+        DataContext = viewModel;
+        InitializeComponent();
+    }
+
+    private void OnPageLoaded(object sender, RoutedEventArgs e) => _viewModel.OnNavigatedTo();
+}